+++ src/datastore/sync_methods/defineResource.js
@@ -101,6 +101,7 @@ var methodsToProxy = [
  */
 function defineResource(definition) {
   var DS = this;
+  var definitions = DS.definitions;
   var IA = DS.errors.IA;
 
   if (DS.utils.isString(definition)) {
@@ -124,9 +125,43 @@ function defineResource(definition) {
   try {
     // Inherit from global defaults
     Resource.prototype = DS.defaults;
-    DS.definitions[definition.name] = new Resource(DS.utils, definition);
+    definitions[definition.name] = new Resource(DS.utils, definition);
 
-    var def = DS.definitions[definition.name];
+    var def = definitions[definition.name];
+
+    // Setup nested parent configuration
+    if (def.relations && def.relations.belongsTo) {
+      DS.utils.forOwn(def.relations.belongsTo, function (relatedModel, modelName) {
+        if (!DS.utils.isArray(relatedModel)) {
+          relatedModel = [relatedModel];
+        }
+        DS.utils.forEach(relatedModel, function (relation) {
+          if (relation.parent) {
+            def.parent = modelName;
+            def.parentKey = relation.localKey;
+          }
+        });
+      });
+    }
+
+    def.getEndpoint = function (attrs, options) {
+      var parent = this.parent;
+      var parentKey = this.parentKey;
+      options = options || {};
+      if (!('nested' in options)) {
+        options.nested = true;
+      }
+      if (parent && parentKey && definitions[parent] && options.nested) {
+        if (DS.utils.isObject(attrs) && parentKey in attrs) {
+          return DS.utils.makePath(definitions[parent].getEndpoint(attrs, options), attrs[parentKey], this.endpoint);
+        } else if ((DS.utils.isNumber(attrs) || DS.utils.isString(attrs)) && DS.get(this.name, attrs) && parentKey in DS.get(this.name, attrs)) {
+          return DS.utils.makePath(definitions[parent].getEndpoint(attrs, options), DS.get(this.name, attrs)[parentKey], this.endpoint);
+        } else if (options && options.parentKey) {
+          return DS.utils.makePath(definitions[parent].getEndpoint(attrs, options), options.parentKey, this.endpoint);
+        }
+      }
+      return this.endpoint;
+    };
 
     // Remove this in v0.11.0 and make a breaking change notice
     // the the `filter` option has been renamed to `defaultFilter`
@@ -227,7 +262,7 @@ function defineResource(definition) {
     return def;
   } catch (err) {
     DS.$log.error(err);
-    delete DS.definitions[definition.name];
+    delete definitions[definition.name];
     delete DS.store[definition.name];
     throw err;
   }